backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==2.26.0
aiosqlite==0.20.0
python-dotenv==1.0.1
